Initial refactor: extract the plugin-agnostic code
