+++ packages/ts-etl/test/unit/crossref/api.test.ts
@@ -1,6 +1,6 @@
 import test from 'ava'
 import { isLeft, isRight } from 'fp-ts/lib/Either'
-import { fetchPublicationByDoi } from '../../../src/plugins/crossref'
+import { makeRoutine } from '../../../src/plugins/crossref'
 import { whenThenResolve } from '../utils'
 import * as cm from '../__fixtures__/crossref'
 
@@ -12,32 +12,25 @@ test('fetchPublicationByDoi: happy-path scenario: a manuscript with one preprint
     cm.mockCrossrefManuscriptWithPreprintResponse,
   )
   whenThenResolve(mocks.worksT.getWorks, { doi: cm.PREPRINT_DOI }, cm.mockCrossrefPreprintResponse)
+  const routine = makeRoutine(mocks.crs)
 
-  const res = await fetchPublicationByDoi(mocks.crs, {}, cm.MANUSCRIPT_DOI)
+  const res = await routine(cm.MANUSCRIPT_DOI)()
 
   if (isLeft(res)) {
-    t.fail(`Got error instead of docmaps: ${res.left}`)
+    t.fail(`Got error instead of steps: ${res.left}`)
     return
   }
 
-  t.is(res.right.length, 1)
-  const dm = res.right[0]
+  const steps = res.right
+  t.is(steps.length, 2)
 
-  // necessary because Typescript doesn't narrow down type of dm just because
-  // test failure guarantees we can't get here
-  if (!dm) {
-    t.fail('impossibly, we couldnt find the first docmap in a list of one')
-    return //necessary
-  }
-
-  t.deepEqual(dm.type, 'docmap')
-  t.is(dm.steps ? Object.keys(dm.steps).length : 0, 2)
-  t.is(dm.steps?.['_:b0']?.inputs.length, 0)
-  t.deepEqual(dm.steps?.['_:b0']?.actions[0]?.outputs[0]?.doi, cm.PREPRINT_DOI)
-  t.deepEqual(dm.steps?.['_:b1']?.inputs[0]?.doi, cm.PREPRINT_DOI)
-  t.deepEqual(dm.steps?.['_:b1']?.actions[0]?.outputs[0]?.doi, cm.MANUSCRIPT_DOI)
+  t.is(steps[0]?.inputs?.length, 0)
+  t.deepEqual(steps[0]?.actions[0]?.outputs[0]?.doi, cm.PREPRINT_DOI)
+  t.deepEqual(steps[1]?.inputs[0]?.doi, cm.PREPRINT_DOI)
+  t.deepEqual(steps[1]?.actions[0]?.outputs[0]?.doi, cm.MANUSCRIPT_DOI)
   //TODO: can write stronger assertions as we learn what this should look like
 })
+
 test('fetchPublicationByDoi: happy-path scenario: a manuscript discovered from its preprint', async (t) => {
   const mocks = cm.CrossrefClientMocks()
   whenThenResolve(
@@ -46,66 +39,24 @@ test('fetchPublicationByDoi: happy-path scenario: a manuscript discovered from i
     cm.mockCrossrefManuscriptWithPreprintResponse,
   )
   whenThenResolve(mocks.worksT.getWorks, { doi: cm.PREPRINT_DOI }, cm.mockCrossrefPreprintResponse)
+  const routine = makeRoutine(mocks.crs)
 
-  const res = await fetchPublicationByDoi(mocks.crs, {}, cm.PREPRINT_DOI)
+  // FIXME:  this makes the call signature look wrong. double-check!
+  const res = await routine(cm.PREPRINT_DOI)()
 
   if (isLeft(res)) {
-    t.fail(`Got error instead of docmaps: ${res.left}`)
+    t.fail(`Got error instead of steps: ${res.left}`)
     return
   }
 
-  t.is(res.right.length, 1)
-  const dm = res.right[0]
+  t.is(res.right.length, 2)
+  const steps = res.right
 
-  // necessary because Typescript doesn't narrow down type of dm just because
-  // test failure guarantees we can't get here
-  if (!dm) {
-    t.fail('impossibly, we couldnt find the first docmap in a list of one')
-    return //necessary
-  }
-
-  t.deepEqual(dm.type, 'docmap')
-  t.is(dm.steps ? Object.keys(dm.steps).length : 0, 2)
-  t.is(dm.steps?.['_:b0']?.inputs.length, 0)
-  t.deepEqual(dm.steps?.['_:b0']?.actions[0]?.outputs[0]?.doi, cm.PREPRINT_DOI)
-  t.deepEqual(dm.steps?.['_:b1']?.inputs[0]?.doi, cm.PREPRINT_DOI)
-  t.deepEqual(dm.steps?.['_:b1']?.actions[0]?.outputs[0]?.doi, cm.MANUSCRIPT_DOI)
+  t.is(steps[0]?.inputs.length, 0)
+  t.deepEqual(steps[0]?.actions[0]?.outputs[0]?.doi, cm.PREPRINT_DOI)
+  t.deepEqual(steps[1]?.inputs[0]?.doi, cm.PREPRINT_DOI)
+  t.deepEqual(steps[1]?.actions[0]?.outputs[0]?.doi, cm.MANUSCRIPT_DOI)
   //TODO: can write stronger assertions as we learn what this should look like
-})
-test('fetchPublicationByDoi: happy-path scenario: publisher is included', async (t) => {
-  const mocks = cm.CrossrefClientMocks()
-  whenThenResolve(
-    mocks.worksT.getWorks,
-    { doi: cm.MANUSCRIPT_DOI },
-    cm.mockCrossrefManuscriptResponse,
-  )
-
-  const publisher = {
-    id: 'my_pub_id',
-    name: 'my_name',
-  }
-  const res = await fetchPublicationByDoi(mocks.crs, publisher, cm.MANUSCRIPT_DOI)
-
-  if (isLeft(res)) {
-    t.fail(`Got error instead of docmaps: ${res.left}`)
-    return
-  }
-
-  t.is(res.right.length, 1)
-  const dm = res.right[0]
-
-  // necessary because Typescript doesn't narrow down type of dm just because
-  // test failure guarantees we can't get here
-  if (!dm) {
-    t.fail('impossibly, we couldnt find the first docmap in a list of one')
-    return //necessary
-  }
-
-  t.deepEqual(dm.type, 'docmap')
-  t.deepEqual(dm.publisher, {
-    id: 'my_pub_id',
-    name: 'my_name',
-  })
 })
 
 test('fetchPublicationByDoi: happy-path scenario: a manuscript with no relations', async (t) => {
@@ -115,27 +66,20 @@ test('fetchPublicationByDoi: happy-path scenario: a manuscript with no relations
     { doi: cm.MANUSCRIPT_DOI },
     cm.mockCrossrefManuscriptResponse,
   )
+  const routine = makeRoutine(mocks.crs)
 
-  const res = await fetchPublicationByDoi(mocks.crs, {}, cm.MANUSCRIPT_DOI)
+  // FIXME:  this makes the call signature look wrong. double-check!
+  const res = await routine(cm.MANUSCRIPT_DOI)()
 
   if (isLeft(res)) {
-    t.fail(`Got error instead of docmaps: ${res.left}`)
+    t.fail(`Got error instead of steps: ${res.left}`)
     return
   }
 
   t.is(res.right.length, 1)
-  const dm = res.right[0]
-
-  // necessary because Typescript doesn't narrow down type of dm just because
-  // test failure guarantees we can't get here
-  if (!dm) {
-    t.fail('impossibly, we couldnt find the first docmap in a list of one')
-    return //necessary
-  }
-
-  t.deepEqual(dm.type, 'docmap')
-  t.deepEqual(dm.steps?.['_:b0']?.inputs.length, 0)
-  t.deepEqual(dm.steps?.['_:b0']?.actions[0]?.outputs[0]?.doi, cm.MANUSCRIPT_DOI)
+  const steps = res.right
+  t.deepEqual(steps[0]?.inputs.length, 0)
+  t.deepEqual(steps[0]?.actions[0]?.outputs[0]?.doi, cm.MANUSCRIPT_DOI)
 })
 
 test('fetchPublicationByDoi: happy-path scenario: a manuscript with 2 reviews and no preprint', async (t) => {
@@ -156,30 +100,24 @@ test('fetchPublicationByDoi: happy-path scenario: a manuscript with 2 reviews an
     { doi: cm.REVIEW_2_DOI },
     cm.mockCrossrefReviewsResponses[1],
   )
+  const routine = makeRoutine(mocks.crs)
 
-  const res = await fetchPublicationByDoi(mocks.crs, {}, cm.MANUSCRIPT_DOI)
+  // FIXME:  this makes the call signature look wrong. double-check!
+  const res = await routine(cm.MANUSCRIPT_DOI)()
 
   if (isLeft(res)) {
-    t.fail(`Got error instead of docmaps: ${res.left}`)
+    t.fail(`Got error instead of steps: ${res.left}`)
     return
   }
 
-  t.is(res.right.length, 1)
-  const dm = res.right[0]
+  const steps = res.right
+  t.is(steps.length, 2)
 
-  // necessary because Typescript doesn't narrow down type of dm just because
-  // test failure guarantees we can't get here
-  if (!dm) {
-    t.fail('impossibly, we couldnt find the first docmap in a list of one')
-    return //necessary
-  }
-
-  t.deepEqual(dm.type, 'docmap')
-  t.deepEqual(dm.steps?.['_:b0']?.inputs.length, 0)
-  t.deepEqual(dm.steps?.['_:b0']?.actions[0]?.outputs[0]?.doi, cm.MANUSCRIPT_DOI)
-  t.deepEqual(dm.steps?.['_:b1']?.inputs[0]?.doi, cm.MANUSCRIPT_DOI)
-  t.deepEqual(dm.steps?.['_:b1']?.actions[0]?.outputs[0]?.doi, cm.REVIEW_1_DOI)
-  t.deepEqual(dm.steps?.['_:b1']?.actions[1]?.outputs[0]?.doi, cm.REVIEW_2_DOI)
+  t.deepEqual(steps[0]?.inputs.length, 0)
+  t.deepEqual(steps[0]?.actions[0]?.outputs[0]?.doi, cm.MANUSCRIPT_DOI)
+  t.deepEqual(steps[1]?.inputs[0]?.doi, cm.MANUSCRIPT_DOI)
+  t.deepEqual(steps[1]?.actions[0]?.outputs[0]?.doi, cm.REVIEW_1_DOI)
+  t.deepEqual(steps[1]?.actions[1]?.outputs[0]?.doi, cm.REVIEW_2_DOI)
 })
 
 test('fetchPublicationByDoi: error case: looking up a crossref work of wrong type', async (t) => {
@@ -189,8 +127,10 @@ test('fetchPublicationByDoi: error case: looking up a crossref work of wrong typ
     { doi: cm.REVIEW_1_DOI },
     cm.mockCrossrefReviewsResponses[0],
   )
+  const routine = makeRoutine(mocks.crs)
 
-  const res = await fetchPublicationByDoi(mocks.crs, {}, cm.REVIEW_1_DOI)
+  // FIXME:  this makes the call signature look wrong. double-check!
+  const res = await routine(cm.REVIEW_1_DOI)()
 
   if (isRight(res)) {
     t.fail(`Got docmaps instead of error: ${res.right}`)
